+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ## [Unreleased]
 
+### Changed
+
+- **(Breaking)** The `setup` function has been removed. Instead, configuration uses the `vim.g.compile_mode`.
+
 ## [3.0.1] - 2024-07-08
 
 ### Fixed